Organize code and folder structure
