+++ config/swagger.js
@@ -19,7 +19,7 @@ const options = {
       },
     ],
   },
-  apis: ['./routes/*.js', './controllers/*.js'],
+  apis: ['./routes/api/*.js', './controllers/*.js'],
 };
 
 const specs = swaggerJsdoc(options);
