+++ routes/api/events.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 
-const { getAllEvents, getEventById, getEventByName, createEvent } = require('../controllers/eventController');
-const { validateEvent } = require('../validators/eventValidator');
-const handleValidationErrors = require('../middleware/validationMiddleware');
+const { getAllEvents, getEventById, getEventByName, createEvent } = require('../../controllers/eventController');
+const { validateEvent } = require('../../validators/eventValidator');
+const handleValidationErrors = require('../../middleware/validationMiddleware');
 
 // https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#schema-object
 /**
