+++ routes/api/users.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { getAllUsers } = require('../controllers/userController');
+const { getAllUsers } = require('../../controllers/userController');
 
 /**
  * @swagger
